+++ softmax.cpp
@@ -8,7 +8,7 @@ double sofLayer::crossEntropyError(const matrix& x, const matrix& t) {
     }
     double sum = 0;
     double delta = 1e-7;
-    for (int i = 0; i < x.col; i++) sum += t.data[0][i] * log(x.data[0][i] + delta);
+    for (size_t i = 0; i < x.col; i++) sum += t.data[0][i] * log(x.data[0][i] + delta);
     double error = (-sum) / x.col;
     return error;
 }
@@ -21,8 +21,8 @@ matrix sofLayer::softmax(const matrix& x) {
     }
     matrix y(1, x.col);
     double sum = 0;
-    for (int i = 0; i < x.col; i++) sum += exp(x.data[0][i]);
-    for (int i = 0; i < x.col; i++) y.data[0][i] = exp(x.data[0][i]) / sum;
+    for (size_t i = 0; i < x.col; i++) sum += exp(x.data[0][i]);
+    for (size_t i = 0; i < x.col; i++) y.data[0][i] = exp(x.data[0][i]) / sum;
     return y;
 }
